integert

+++ apps/partner/strategy.py
@@ -31,7 +31,7 @@ class AbchauzIndiaFixedRateTax(FixedRateTax):
         tax = (stockrecord.price_excl_tax * rate).quantize(exponent)
         return TaxInclusiveSellingOrientedFixedPrice(
             currency=stockrecord.price_currency,
-            incl_tax=math.floor(stockrecord.price_excl_tax + tax),
+            incl_tax=int(math.floor(stockrecord.price_excl_tax + tax)),
             tax=tax)
 
     def parent_pricing_policy(self, product, children_stock):
@@ -47,7 +47,7 @@ class AbchauzIndiaFixedRateTax(FixedRateTax):
 
         return TaxInclusiveSellingOrientedFixedPrice(
             currency=stockrecord.price_currency,
-            incl_tax=math.floor(stockrecord.price_excl_tax + tax),
+            incl_tax=int(math.floor(stockrecord.price_excl_tax + tax)),
             tax=tax)
